requirements.txt
vecs
flask
llama-index
requests
html2text
python-dotenv
